# deprecated/cosmosdemo/graph-query-api/openapi/templates/graph.yaml
openapi: "3.0.3"
info:
  title: Graph Query API
  version: "0.5.0"
  description: |
    Micro-API for executing graph queries against the topology database.
    Used by Foundry agents via OpenApiTool.
# NOTE: Template placeholders ({base_url}, {graph_name}, {query_language_description})
# are substituted at runtime by agent_provisioner.py.
servers:
  - url: "{base_url}"
    description: Deployed Container App

paths:
  /query/graph:
    post:
      operationId: query_graph
      summary: Execute a query against the topology graph
      description: |
        {query_language_description}
        Returns columns and data rows from the network topology ontology.
      parameters:
        - name: X-Graph
          in: header
          required: true
          schema:
            type: string
            enum: ["{graph_name}"]
          description: |
            The graph name to route this request to. Always use the value shown.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: |
                    {query_language_description}
      responses:
        "200":
          description: Query results (check 'error' field for query failures)
          content:
            application/json:
              schema:
                type: object
                properties:
                  columns:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        type:
                          type: string
                  data:
                    type: array
                    items:
                      type: object
                  error:
                    type: string
                    nullable: true
                    description: |
                      If present, the query failed. Contains the error message.
                      Read the message, fix your query, and retry.